+++ app/Models/Category.php
@@ -15,4 +15,9 @@ class Category extends Model
     {
         return $this->hasMany(Topic::class);
     }
+
+    public function getRouteKeyName()
+    {
+        return 'slug';
+    }
 }
